requirements.txt
langchain
langgraph
langchain-openai
langchain-community
azure-identity
openai
asyncpg
pgvector
pydantic
pydantic-settings
pyyaml
httpx
jinja2
python-dotenv
tiktoken
numpy
pandas
jamo>=0.4.1
rapidfuzz>=3.0.0
python-mecab-ko>=1.0.0
g2pk>=0.9.4
scikit-learn>=1.0.0